Move themes to lib/
